requirements.txt
flask
google-cloud-language
google-cloud-speech
pydub
moviepy
werkzeug